Final code to be pushed to the master branch
This commit contains code changes to derive scaling factors applied on precipitation
fluxes to ensure water balance in ocean model. The idea here is based on the globally
averaged water balance equation:

<dS/dt> = <P> - <E> - <R>
where <dS/dt> denotes the change of ocean water storage, <P>, <E> and <R>
refer to the precipitation, evaporation and river runoff fluxes.

when <dS/dt> /= 0, the precipitation factor f is introduced and applied on
<P> so that
0 = f*<P> - <E> - <R>. where the precipitation factor f is estimated as

f = 1.0 - <dS/dt> / <P>. Here, the <dS/dt> is defined as

<dS/dt> = (SSH(final) - SSH(initial)) * rho_sw / second_in_year

The code in this commit is to add the above calculation process as a
module in ocean model so that the f is dynamically diagosed at the end
of each year (by default) and then applied as a scaling factors on the
precipitation fluxes in the following year simulation.

The code has been verified with 10-year simulations with two different
configurations with coupled historical data atmosphere forcing, and
results indicate that the code worked as expected. The restart workflow
was also tested and verified sucessfully.