There are two diagnostic output fields that should be optional.
The column integrated modal values of the deposition flux
of aqueous so4 and the deposition flux of aqueous h2so4
should be optional.

This will be a non-BFB change sinde these values were
always output before and now the default is not to
create these fields and avoid the calculation.